Store anonymous user and pass the users to frontend.
We make sure that an anonymous user is recorded in
our DB when a user joins the "admin:active_users" and
"room:#{room_id}" topics. Both are possibly the first
entry points to our app for our user.

Our anonymous users are passed as a 'lobby_list'
to the frontend when an admin joins the "admin:active_users"
topic so the sidebar can be populated with these users.
Now, our admin can see users they've chatted with over
the week/month/year!

The "lobby_list" event is only meant for admins. To
ensure that only admins receive payloads for this event,
we `intercept/1` it and use `handle_out/2` to filter
only the admins with the same public key as the user
will receive the event.

Note that we extract public functions that are used or
will be used in both AdminChannel and RoomChannel.
Keeping our channels clean of public functions that
are not callbacks is a good way to keep them clean
and maintainable.