Record when admin views a chatroom and send more events.
We expect "previousRoom" and "nextRoom" in the payload
when a user joins a room topic. We use this to record
a timestamp of when admin has viewed both rooms. This
means we record when an admin both leaves a previous
room and joins a new room.

We then publish two more events every time an
anonymous user sends a message. The "lobby_list" event
for updating the user's last message sent and timestamp
and the "notifications" event for triggering browser
notifications. We'll get into more detail on that
once we work on the frontend.