Ensure that anonymous user is saved.
When an anonymous user joins admin:active_users topic, we
make sure that user is saved to our DB. Also, we no longer
track the presence of admins to simplify our code especially
in the frontend (no more filtering out admin's id).

We also add a simple validation on the params passed on socket
connect. We raise an error if both id and uuid params are
empty. That way, code that gets executed in channels can expect
that a socket.assigns will always have a non-empty user_id
or uuid. This helps to simplify our code too and informs
the frontend that an id or uuid param is expected.