Refactor to use @derive for Poison.
Defining a @derive module attribute is a convenient
way to customize/instruct how Protocols treat our struct.
We use @derive for Poison.Encoder in our Message and
AnonymousUser struct so that Poison knows which fields
to include when it encodes it to JSON. This saves us
from writing a bunch of code for specifying which
fields to include in our JSON version of our structs.

Apart from specifying the fields for our JSON payloads,
we also pass the most recently active users as lobby_list
once admin joins the admin:active_users topic. This
allows the frontend to load an initial list of users
in the sidebar for admin.