Update lobby_list for admins.
We broadcast to admins every time user connects to our
app via the lobby_list event. Our frontend will then be
able to listen to this event and handle updates accordingly.
This event will get triggered every time a user joins
the admin:active_users topic so it will be up to the
frontend to check whether a user is already on the
list.